+++ changelog.md
@@ -13,6 +13,7 @@
     * Control + P: Play/pause.
     * control+Shift+P: Play all.
 * Fixed an error in the audio player that was skipping the first track if you were in the last song and pressed "play next" in the menu bar or via the keystroke.
+* Chats with unread messages will be placed at the top of the chats section. When a chat buffer receives a new message, socializer will move  the buffer to the first position in the chats list. This should make easier for everyone to determine which chats contain unread items. ([#24,](https://code.manuelcortez.net/manuelcortez/socializer/issues/24))
 
 ## Changes in version 0.18 (21.01.2019)